Updated openstack/openstack Project: openstack-infra/system-config  7daf066b65f482e007ca5f7d026264b890c276f2
Add #openstack-shade to managed channel list

Although shade is an infra project, it has a bunch of development
interest from sources that are not Infra related, such as Ansible and
the AppDev Work Group. Unfortunately for many of those folks,
the Infra is a very busy, so discussion often happens in places that is
not logged. Let's go ahead and make sure that such discussions are
logged and whatnot.

Depends-On: Ie9d2d5781cbc5a7e072bd7fbfde09eb47078cf04
Change-Id: Ic16d6ac6fae216efea815bacc99a0021d9935d1b