Updated openstack/openstack Project: openstack/tempest  970a87b985c0059ea1aa17708a6b56732b718f92
Stop using subprocess for testr init in tempest init

This commit switches the call to testr init from using subprocess to
testr's python run_argv interface. (which is essentially just calling
it via subprocess) The issue with using subprocess directly is that it
assumes testr is in the PATH you're running from. But, in the case of
using a venv this isn't always the case.

Change-Id: I2bab70b0e9fb2ccdc340f48d95340ff7ccdc1364
Closes-Bug: #1633182