Updated openstack/openstack Project: openstack-infra/project-config  33ca876e9917d39ffb1630d172db5a6c68a806c1
add tripleo non-voting job into osc and osc-lib gates

With the motivation of adding more diversity in how do we test
openstackclient and osc-lib, this patch will add a tripleo job,
non-voting.

The idea is to bring more feedback when testing patches in these
projects and detect what failures we could have outside devstack.

The jobs are not voting. We don't expect OSC team to open TripleO job
logs. The idea is to let TripleO folks know when a job is red so they
can investigate and tell if something critical is broken in the patch.

Change-Id: I7e3dcf11c66be872d1895d299cf9e61d85996c01