Updated openstack/openstack Project: openstack/python-keystoneclient  3c8f1276a2f874f3fcaf7db08761d79bd013cf28
Use exceptions from Keystoneauth

As keystoneclient and other services rely more on keystoneauth we should
assume that keystoneauth is our base auth library, not keystoneclient
and start to default to the objects provided from there. This will make
it easier to remove these objects when the time comes.

For the session independant parts of keystoneclient we should use the
exception names as provided by keystoneauth instead of the aliases in
keystoneclient.

Change-Id: Ic513046f8398a76c244e145d6cc3117cdf6bb4cd