Updated openstack/openstack Project: openstack/python-barbicanclient  a998f7576a6e0263f802b19bd88b51d4890c58ea
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed). If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit
Tests fail. The following patch fixes this issue.

Change-Id: I5209e5889cfe62c66dc675c00e3ccaf22410f637
Partial-Bug: #1259292