Updated openstack/openstack Project: openstack/keystone  f98e8fd0db45936e95aab6b44a6a9c7d2cbb6a95
Invalidate trust when the trustor or trustee is deleted

The trust without a valid trustee, trustor is useless and will
no longer be active since the id of user is a random number and
only assigned when it created.

The patch delete trust if the related trustee or trustor is
deleted if the user is maintained by keystone.

Change-Id: I67dac6b7bac8cb94575ceda4a3277847a2bcc2d8
Partial-Bug: #1622310