Updated openstack/openstack Project: openstack/horizon  0409080e7b93a91323bbb69e1811ef79b17c4e6d
i18n: The same use of args with ugettext_lazy

Different use of args in ungettext_lazy causes
error on import job from translation infrastructure
to horizon repository.

The use of variables in singular and plural strings
needs to be same. This commit also adjusts the string
with ugettext_lazy() as other strings are dealt with.

Change-Id: I9a836178b2d615504950545654242c0a4c196723
Closes-Bug: #1630507