Updated openstack/openstack Project: openstack-infra/storyboard  53c378ba93d31a5c3f3b5134c9670231a62debc7
Add users/self API

Add an API method to return the User record for the currently
logged-in user.

Without this, the only way to find out the user_id of the current
user is to perform an OpenID authentication.  This allows users
with only a valid access token to know their user id.

Change-Id: I0284b4bfcef731a9433b47cc165009518fc67a99