Updated openstack/openstack Project: openstack/cinder  d9d1fadf62591fe4aacd6b61ebbf83c40da8d4bf
Fix HNAS Driver KeyError exception in volume manage

Hitachi cinder driver throws a KeyError when one tries to
manage a volume without specifying a pool and without having
a service label configured as 'default'. This affects both
NFS and iSCSI drivers.

Adding a check to verify if a 'default' HNAS service type
exists when a volume type without service_label is passed
in volume manage and giving a proper exception message
when it fails instead of a KeyError exception.

Change-Id: I97251de46ee6525e13c8c0130af1d390959b92ff
Closes-Bug: #1628466