Updated openstack/openstack Project: openstack/swift  a79d8508df493d5744b262e1d1830782e32dbd04
Make Content-Disposition support `inline; filename=...` format.

A client can process a object contents immediately when it has knowledge,
if it doesn't, it stores the content to local with the filename that
assigned in `Content-Disposition`.

According to https://tools.ietf.org/html/rfc2183, `filename` can be use
whatever `inline` or `attachment` (disposition-type) in the disposition
string.

Change-Id: Iba94aedfad06c1dc8bd7be5eb3c73b33fb8d5198