Updated openstack/openstack Project: openstack/tempest  db3acad95c2b90408d389bced9232817f9bce765
Revert an accidental error modification of config item help message

The help message of config.compute.fixed_network_name is changed in
Ied68c0326cf891a4e6eb3db51dc5f43241e4b8d2 accidiently which is unwanted,
so this is to revert it.
Besides, this is to remove some unwanted words.

Change-Id: Ia4979772faf33d90f9544c510f4c421c4fa0097f