Updated openstack/openstack Project: openstack/heat  3a062fdf22f2edeab8045c08e5da1dfeddc9616b
Add new hot function str_replace_strict

In many cases, a user would rather see an error result if a
str_replace param is not substituted in the template, rather than the
function silently doing nothing.

Since str_replace is set in its ways, introduce a new function
str_replace_strict which behaves identically to str_replace except
that a ValueError is raised if any of the param's are not found
in the template.

Change-Id: I8b8c69bb49dfeb74e05af4871602c20493b081eb