Updated openstack/openstack Project: openstack/python-troveclient  a06ad17715d06ad95d52430e489929d9d956adc6
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I6af6b4e92f0953f673ea1a5c0f2162662cb14584