Updated openstack/openstack Project: openstack/ironic  022f624cfc231c7198bc0161ddd73ab364bd3503
Clean up provision ports when reattempting deploy

If the provisioning fails at boot.prepare_ramdisk stage, as after
that provision ports are not cleaned up, when retrying, neutron
network driver won't be able to create ports, complaining about
duplicated MAC addresses. The only way out of this is doing
set-provision-state deleted and then active again, but then
instance_info will be lost, which is inconvenient in case of
standalone usage.

Closes-bug: #1607394
Change-Id: Ifacbeb307619a9a7883ab1ac8fce3ff07e46787c