Updated openstack/openstack Project: openstack-dev/pbr  e8613428c15d5943086879fd942f86ec84b6614c
Fix handling of old git log output

Older versions of git produce log output (using --format specifiers)
that doesn't prefix tags with a "tag:" prefix, though they do using
the --decorate log format.  This was a bug in git that was fixed, but
the older git is still in the wild.  This change combines both the new
--format specifier method with the --decorate method to find tags even
if used with the old, buggy git, while (hopefully) avoiding
identifying non-tags as tags.

This is based on I6b557687f2428a528a8aa7f76e84797162488887 but tries to
use a single format string that works across all versions of git.

Because the format string has changed, some of the test data has had to
change to reflect the new format.

Change-Id: I58a1b12e0f4054fd25bddc77fabb6c89239561ff
Closes-Bug: #1570458
Co-Authored-By: James Polley <jp@jamezpolley.com>