Updated openstack/openstack Project: openstack/glance_store  9b63cb5d00b26b6d8bf105ae306094775cacd8dc
Refactor get_manager_for_store in an OO manner

get_manager_for_store is a weird function that switches based on the
class type of the first parameter. This is an odd throw back to pre
object orientated days where the object defines what it wants. Refactor
it to put the class on the object.

Change-Id: I0bca2607267aef3bda720cdfbbbe0e5a8093a20d