Updated openstack/openstack Project: openstack-infra/devstack-gate  01b55c8412f463b2da0569e43dfcad5614da8559
Reorganize PROJECTS list to show repos to remove

We want to reduce the size of the PROJECTS list by default to more
closely match things done by a default devstack + tempest test run. The
reason for that is every project in the PROJECTS list has to be cloned,
updated, and checked out to the proper ref as part of setup. This can be
slow so reduce the total number of projects that need this by default.

In this change we split the list into two halves, the first is the half
that should remain and the second is the half that we should work to
remove. Hopefully this also discourages new changes to add projects to
this list.

Change-Id: Iaec3d0ad54879c22cd14907c15c8352ba1c63ec8