Updated openstack/openstack Project: openstack/ceilometermiddleware  163b2060fae1492c04eed52e67ce61c3740c710a
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I53d883cdcd947b2c0c9fb55330b7ee70430f25cc