Updated openstack/openstack Project: openstack/training-guides  e85b57787225a346069f8d1e0b0dfc2bb1e9726b
[upstream] Update of Events Module

With the recent updates to OpenStack events(i.e. the advent of the PTG),
this module needed to be reworked. There are OpenStackDays events now
as well that weren't previously mentioned. This patch adds this
new information to this module. It also renames the module to 'OpenStack
Events' because there are many more events than just the Design Summit that
students should be aware of.

Change-Id: I0ecbde9828e09e0781cbdaf23101da03f785d4e6