Updated openstack/openstack Project: openstack/neutron  d2a28a8c73a07cbccd79ce82c0738e629b4134a8
Only emit deprecation warning when deprecated value read

If providers is not found in the CONF.service_providers.service_provider
it looks them up from implicit conf files. If nothing is defined there
it still emits a deprecation warning.

Maybe a warning should be emitted if this value is empty, however the
deprecation warning that is provided is incorrect and misleading.

Change-Id: I3166a445b69d854b294843814a081e74c1aadb3b
Closes-Bug: #1633293