Updated openstack/openstack Project: openstack/keystone  7a78ac0999f6ea0a8d34247b22859e25ee7f2a13
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I371da7d4185a14a59d959b02b1301a71f5fb2551