Updated openstack/openstack Project: openstack/nova  7bc78d06388af00e43539ed3ef45ac335ce6ec9c
Explicitly name commands target environments

Naming envs like this is breaking the current
oslo periodic day job; a simple way to unbreak this
is to name them explicitly (which appears to be
more common).

So this changes the command listing to do just that
as well as lists py35 in the envlist (because its
a valid target).

Change-Id: I2a162604b60df8f18ff53069c6bf99cfecc4217e