Updated openstack/openstack Project: openstack/python-designateclient  9237e1faf3d60cf04b72af6fc85b364fb8b05a4c
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I477eaf74541c65ad7ab54fa8c2412c8b9c74c684