Updated openstack/openstack Project: openstack/governance  0623fe19d16d24586b5b2a9638ca8a873b5bcc93
Add "Contributing Is Our Currency" to principles doc

This commit adds a new principle to the principles doc titled
contributing is our currency. The goal here is to express that while
we have a hierarchical leadership structure the only way change gets
implemented is by people actually contributing and doing the work.

Change-Id: I876f95ccbbf71f4f512275af110690888449e093