Updated openstack/openstack Project: openstack/neutron-fwaas  7e6e20681d14612c0e7a1001c07ebaa6348ce1e1
Fix UnboundLocalError during update firewall group.

In l3 agent fwaas v2 logic, it may cause the local varible 'fw_ports'
not defined, so we should set it.

This patch fix it when an new collection of ports_for_fwg, as this time
it will store the 'add-ports' list. Then re-defining 'fw_ports'.

Change-Id: Idb9d7987bed74b6e78ae153610e2c351c2c389f0
Closes-Bug: #1634114