Updated openstack/openstack Project: openstack-infra/zuul  380103662e13f96b506d4a405bee8387397bb2ff
Split playbook into vars, pre-playbook and playbook

In order to get to the point where playbooks that people write for tests
are playbooks that they could conceivably also use outside of the zuul
context, we need to remove the need for zuul-specific things in the main
playbook.

Add a pre-playbook that runs before the playbook and runs the things
that are not tied to current JJB content - namely setting up the logger
and prepping directories.

Move the SUCCESS/FAILURE message to the post-playbook.

Extract the injected variables into a variables file and add a
-e@vars.yaml option to the playbook invocation. This provides variables
in a known namespace. Obviously there is still an exercise in how a user
might write a playbook that wants to consume those variables in some
way.

Change-Id: Ie5ec6ec65a03ceea9afc3ac59df73cb28f5ca4dd