Updated openstack/openstack Project: openstack-infra/zuul  b27b7e1c5d2c4b2125fa16386b0227cd83fc9916
Enable pipelining for ansible-playbook

From the manual:

  Enabling pipelining reduces the number of SSH operations required to
  execute a module on the remote server, by executing many ansible
  modules without actual file transfer. This can result in a very
  significant performance improvement when enabled, however when using
  “sudo:” operations you must first disable ‘requiretty’ in
  /etc/sudoers on all managed hosts.

Basically on local testing, there is a speed improvement.  However, I
believe the better reason to enable this is to reduce the number of
SSH transactions we preform on our workers. In doing this we reduce
our potential chances for SSH connection issue.

However, it also appears async operations do not use this setting
simply because of async works.

Change-Id: Ib224fbf1fed19be3ce7db4da0c466e3d11acc365
Signed-off-by: Paul Belanger <pabelanger@redhat.com>