Updated openstack/openstack Project: openstack-infra/shade  7d02b5d2092a697bff486f42ce7c52cf26527cd4
Add abililty to find floating IP network by subnet

If the cloud has a hidden public network but a visible router, it's not
possible to infer network info from metadata, but it is possible to know
that the network that the floating IP should be requested on is the one
that is in the router's external_gateway_info field.

Wrap the search in local caching and a lock to prevernt thundering herd.
It's not a thing that will need to be inferred more than once, but has
the potential to be hit on every server creation for clouds that need
this.

Change-Id: Ic98b6059e0ca5d0f1c8e6db16cc89a352c6fbbbd