Updated openstack/openstack Project: openstack/neutron-vpnaas  c150c6063acd41453e1b4b7aa6cd7a63d2e83d36
Remove FWaaS dependency in VPNaaS devstack plugin

Right now if the FWaaS configuration file does not exist, a devstack
trying to start VPNaaS may fail[1].  This is unnecessary; since FWaaS is
now using the L3 agent extension code, it plugs in that way.  It should
not interfere at all with VPNaaS.  This change removes the FWaaS
dependency in the devstack plugin.

[1] http://paste.openstack.org/show/584360/

Change-Id: I28e48ab161e7991a24fec181e4ec7378888f9149