Updated openstack/openstack Project: openstack-infra/elastic-recheck  539e0905a4a644d7455101ffc588d1fa12a08525
Remove query for pillow 3.0.0 bug 1501995

This was fixed awhile back and just started spiking
again because of some os ansible projects that aren't
using upper-constraints for functional test jobs.

Since that's a different failure, this is misleading
and is listed as the top check queue failure, so let's
remove it.

Change-Id: I459ab2eb2bc8864e762c8360dc46ebcac93b9a0f