Updated openstack/openstack Project: openstack-infra/project-config  fd56be5e44cef5e12f5f4ff6bb4f9f9452ab4a05
Properly find tarball for NPM packages we publish

For js-openstack-lib project, the tarball name is
"openstack-lib-$VER.tgz" and not "js-openstack-lib-$VER.tgz" as we
want to publish the packge without the "js-" prefix.
In order to publish properly, this patch introduce a less restrictive
regex to find the tarball name to publish.

Change-Id: Iffbd40701960b754184eda438436b56b8d87cefa