Updated openstack/openstack Project: openstack/python-keystoneclient  1c48bbc6fcef469eb2f842152bc013572d032ecb
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I008cbed4454d5529373badcd14a16472068c1a7d