Updated openstack/openstack Project: openstack/tempest  628c5bba87d27320f91eb30ae273f7c8dd5e9224
Allow for wait in _restore_password

The _restore_password method within the
test_user_update_own_password() tests attempts to authenticate right
after changing a password. This patch makes it so that we wait one
second before attempt the authentication call, similar to the pattern
already established in the pre-existing test.

Change-Id: Ia51f28a70ae099f1ec93851d271db8556aced836