Updated openstack/openstack Project: openstack/nova-specs  462ba34be936c077a4dcc295fe84d81edffcce3c
Detach and attach boot volumes

It is sometimes useful for a cloud user to be able to
detach and attach the boot volume of an instance that
is not running. Currently Nova does not allow this and
some operations assume it does not happen. This spec
proposes allowing the detach and attach of boot volumes
in the shelved_offloaded state and adding safeguards to
ensure it is safe.

blueprint detach-boot-volume

Change-Id: I07a1e8eae700f35bda31998e6b01e09bdee17ba6
Previously-approved: Newton