Updated openstack/openstack Project: openstack/tempest  707a2cf0d25314fa5cca11cd678fa2059161ccea
Image cleanup is missing in test_create_delete_image

In test_create_delete_image, there is no addCleanup after
self.client.create_image, so if anything goes wrong, the
following self.client.delete_image may not be executed and
the image will be left over improperly. This is to fix it.

Change-Id: Ib54758c0d29efd4f49d470a665713878c0576f10