Updated openstack/openstack Project: openstack-infra/project-config  54670961f6f927d0cd709de823be09bf325ffb70
Ensure a diff does not cause a non-zero return code

Currently periodic jobs for oslo are failing since
tox was thinking the oslo-from-master.sh from master
script was dying. It wasn't actually, this should
fix that by making sure the diff command doesn't return
non-zero (we expect a difference).

Also show the post-alteration tox.ini, so that its
easier to see if a change to it is invalid (especially
when there are some further issue).

Change-Id: I6a73bab746f22e5ae749dc798a189c9f255cddda