Updated openstack/openstack Project: openstack/cinder  ac107a54b800656590edae6eb066ac7b16c664f2
Return HTTPBadRequest instead of HTTPNotFound

If you pass non-existing volume-type or service to the volume manage
API, then it returns 404 HTTPNotFound error.

According to the api guidelines [1] it should return 400
HTTPBadRequest.

[1] https://github.com/openstack/api-wg/blob/master/guidelines/http.rst#failure-code-clarifications

APIImpact: For volume manage API, it will return 400 error instead
of 404 if user passes non-existing volume-type or service in the
request.

Closes-Bug: #1632727
Change-Id: Iffa1824454c765d8cb732e93503420bafca17f08