Updated openstack/openstack Project: openstack/nova  d44c4862e9938bab29260170b4a4a5077e3487ee
tests: Adding functional tests to cover VM creation with sriov

Adding HostPciSRIOVDevicesInfo class to fakelibvirt module
in order to dynamically generate host's SR-IOV devices info.

Adding functional tests to cover the creation of instances with
attached sriov devices (virtual and physical).
Verifying that VFs cannot be allocated when it's parent PF is allocated
and vice-versa, when a child VF is allocated, the parent PF should not
be available.

Change-Id: Ib9f341b0f26d48939d3305f6575f2689682c2685