Updated openstack/openstack Project: openstack/nova  36a37ade2c1596f29adc3b3ca18863abff28b563
Changed the name of the standard resource classes

Originally, the declared names for the resource classes defined in
objects/fields.py were an enum, as the original design assumed that any
future resource classes would be added as a software upgrade release. As
is the convention, this enum was given the name 'ALL'. With the
introduction of custom resource classes that can be added by admins, the
defined classes are no longer an enum, so the name 'ALL' is misleading.
This patch renames it to the more accurate 'STANDARD'.

Change-Id: I251a6faa8c284eb91d190b47108664e67d824425
Blueprint: custom-resource-classes