Updated openstack/openstack Project: openstack/oslo.utils  d0cd71c1fcea97bbe29e5923756a0f5260683ab7
Add threading<->eventlet compatible Event

This creates an object that is backed by either a threading or
eventlet Event, depending on whether eventlet is available.  In the
eventlet case, the object wraps the eventlet.event.Event object to
provide an API that matches that of threading.Event.  This allows the
user to be ignorant of which implementation is backing the object and
to use both interchangeably.

Change-Id: Id33c9f8c17102ba1fe24c12b053c336b6d265501
Related-Bug: #1518430