Updated openstack/openstack Project: openstack/tripleo-heat-templates  9b3aba0309af54991d3e24b0ca147426a5b815a5
Use correct password for keystone bootstrap

In upstream puppet-keystone, the boostrap process should use an admin
password not the admin token for the bootstrapping of keystone. The
admin password option is being added to the upstream class so we will
need to provide it to properly have keystone bootstrapped.

Change-Id: Icab4b0cb70d6caf2f2792c4fe730f060b807fbc1
Depends-On: I7a706d93b43ec025bdb4b29667f64ff2f7dd52a0
Related-Bug: #1621959