Updated openstack/openstack Project: openstack/tempest  75f5a6b6ae575ec8bda6e64b77d94425a753d0b7
Merge the separated link lines on image client

On current docstring of service clients, every link to API
reference is separated to 2 lines. That is unreadable on the
code and the documentation which is created from the docstring
also will be unreadablebecause of the separated link lines.

This patch merges the separated lines into a single line for the
readability.

Change-Id: I1d227163dd21706583a0b771e2e6638b70804f05