Updated openstack/openstack Project: openstack-infra/system-config  1dfddad1150948893c84515871a63b4a000323a6
Add ErrorDocument for docs.o.o

We now have an errordocument, enable using it for 404 errors.

The content is served from openstack-manuals.

Alternative would be to use .htaccess but then the page would be
available on both old and new pages - and the content is specific to the
new AFS based page. Therefore, I choose to use ErrorDocument here.

Change-Id: Iefa7c4453f05071dd990a5ae8a29d2648bb797d9