Updated openstack/openstack Project: openstack-infra/storyboard  4aed46c18926a28788bff991fbe243a41029a85b
Add filters to the /v1/stories/search endpoint

This commit adds almost the same set of parameters that the /v1/stories
endpoint supports to the /v1/stories/search endpoint. This allows one
to obtain some search results and then filter them more easily than
before, as well as avoiding a bug when trying to naively search for
stories without removing the default criteria in the web ui.

The title and description filters are not added, as they don't make
much sense in the context of a fulltext search.

This commit introduces some code duplication, and long term it would
probably be best to merge the two endpoints into one and do the fulltext
search iff the q parameter is given, rather than trying to maintain
two functions with the same set of parameters for no reason.

Change-Id: Ibbe1ed38094fb1d02919cf97ed311cf79af049bb