Updated openstack/openstack Project: openstack/trove  0b8f01a6a76f43b02b79352793bb92b44805ae71
Make trovestack able to run in the gate context

These changes are to make trovestack able to run in the dsvm-gate
context. The fundamental issue we're addressing is that the old
trove-integration was cloned into a writable area but in the current
(trove context) way of doing things, it is a read-only area. Therefore
you can't put the reports directory into the same relative place that
trove-integration would have done.

Change-Id: I449abf5a4ca574a2e9f5ddb1c8be507f5c5429a7