Updated openstack/openstack Project: openstack/neutron  2b9d9e3b9e9e9133b90013114d73356742844ef7
Refactoring config options for cache_utils opts

Refactoring neutron cache_utils config opts to be in neutron/conf so that all
the configuration options for cache_utils reside in a centralized location.
This simplifies the process of looking up the cache_utils config opts and
provides an easy way to import.

Change-Id: I722a2fa7d72c62e6d878d0137f899d09a9ad64ae
Partial-Bug: #1563069