Updated openstack/openstack Project: openstack/nova  7e36553fa16e599b65abb2ad1f032fdfc08cf1c8
Add missing compat routine for Usage object

We added a v1.1 to Usage without a compat routine to go with it. It's not really
a big deal as we're not sending this over RPC anywhere yet that I know of, but
we should have this for the sake of completeness.

Change-Id: Ia50448741b29e1c33e4988a9476418b6ba835983